Add files via upload
We will be using a dataset from a US-based organization called Yelp, which provides a platform for users to provide reviews and rate their interactions with a variety of organizations – businesses, restaurants, health clubs, hospitals, local governmental offices, charitable organizations, etc. Yelp has made a portion of this data available for personal, educational, and academic purposes.

You will be asked a series of questions regarding the data to help you profile and better understand the data in the first part of the assignment. Once you have answered each question, you will come up with your own question for analysis and will prepare a dataset for the analysis you choose to do in the second part of the assignment.